Revert "Api error return response changes with some api bug fixes"

See merge request Pankaj-Baranwal/webapp!11

+++ app/Http/Controllers/LearnWithYoutubeController.php
@@ -331,7 +331,7 @@ class LearnWithYoutubeController extends Controller
             return $this->apiResponse->sendResponse(200, 'New Category added', $category);
         } catch (\Exception $e) {
             DB::rollback();
-            throw new HttpException(500, $e->getMessage());
+            return $this->apiResponse->sendResponse(500, $e->getMessage(), $e->getTraceAsString());
         }
     }
 
@@ -451,19 +451,26 @@ class LearnWithYoutubeController extends Controller
         }
 
         try {
-            $category = Category::find($request->category_id);
-            if ($category->user_id == Auth::user()->id) {
-                $success = $this->deleteAllCategory($category->id);
-                Category::whereIn('id', $success)->delete();
 
+            $category = Category::find($request->category_id);
+            if (is_null($category)) {
                 DB::commit();
-                return $this->apiResponse->sendResponse(200, 'Category deleted successfully', null);
-            } else {
-                return $this->apiResponse->sendResponse(401, 'Unauthorized user can not delete category', null);
+                return $this->apiResponse->sendResponse(201, 'Category Not Found', null);
+            }else {
+                if ($category->user_id == Auth::user()->id) {
+                    $success = $this->deleteAllCategory($category->id);
+                    $learning_paths = LearningPath::whereIn('category_id',$success)->delete();
+                    Category::whereIn('id', $success)->delete();
+
+                    DB::commit();
+                    return $this->apiResponse->sendResponse(200, 'Category deleted successfully', null);
+                } else {
+                    return $this->apiResponse->sendResponse(401, 'Unauthorized user can not delete category', null);
+                }
             }
         } catch (\Exception $e) {
             DB::rollback();
-            throw new HttpException(500, $e->getMessage());
+            return $this->apiResponse->sendResponse(500, $e->getMessage(), $e->getTraceAsString());
         }
     }
 
@@ -502,6 +509,7 @@ class LearnWithYoutubeController extends Controller
         array_walk_recursive($array, function ($a) use (&$return) {
             $return[] = $a;
         });
+       
         return $return;
     }
 
@@ -625,7 +633,7 @@ class LearnWithYoutubeController extends Controller
             }
         } catch (\Exception $e) {
             DB::rollback();
-            throw new HttpException(500, $e->getMessage());
+            return $this->apiResponse->sendResponse(500, $e->getMessage(), $e->getTraceAsString());
         }
     }
 
@@ -633,7 +641,7 @@ class LearnWithYoutubeController extends Controller
     {
         DB::beginTransaction();
         try {
-             $user_id = 1;
+            $user_id = 1;
             //$user_id = Auth::user()->id;
             $getHistory = Video::select('*')
                 ->with('duration_history:video_id,start_time,end_time')
@@ -648,7 +656,7 @@ class LearnWithYoutubeController extends Controller
 //            }
         } catch (\Exception $e) {
             DB::rollback();
-            throw new HttpException(500, $e->getMessage());
+            return $this->apiResponse->sendResponse(500, $e->getMessage(), $e->getTraceAsString());
         }
     }
 
@@ -723,7 +731,7 @@ class LearnWithYoutubeController extends Controller
             }
         } catch (\Exception $e) {
             DB::rollback();
-            throw new HttpException(500, $e->getMessage());
+            return $this->apiResponse->sendResponse(500, $e->getMessage(), $e->getTraceAsString());
         }
     }
 
@@ -751,7 +759,7 @@ class LearnWithYoutubeController extends Controller
             }
         } catch (\Exception $e) {
             DB::rollback();
-            throw new HttpException(500, $e->getMessage());
+            return $this->apiResponse->sendResponse(500, $e->getMessage(), $e->getTraceAsString());
         }
     }
 
@@ -780,7 +788,7 @@ class LearnWithYoutubeController extends Controller
             }
         } catch (\Exception $e) {
             DB::rollback();
-            throw new HttpException(500, $e->getMessage());
+            return $this->apiResponse->sendResponse(500, $e->getMessage(), $e->getTraceAsString());
         }
     }
 
@@ -816,7 +824,7 @@ class LearnWithYoutubeController extends Controller
             }
         } catch (\Exception $e) {
             DB::rollback();
-            throw new HttpException(500, $e->getMessage());
+            return $this->apiResponse->sendResponse(500, $e->getMessage(), $e->getTraceAsString());
         }
     }
 
@@ -876,7 +884,7 @@ class LearnWithYoutubeController extends Controller
             }
         } catch (\Exception $e) {
             DB::rollback();
-            throw new HttpException(500, $e->getMessage());
+            return $this->apiResponse->sendResponse(500, $e->getMessage(), $e->getTraceAsString());
         }
     }
